refactor(ff-probe): split info.rs into probe/builder.rs + probe/probe_inner.rs
Follow the standard module layout used by ff-decode and ff-encode:
- probe/mod.rs: re-exports only
- probe/builder.rs: safe public fn open() with file-system precondition checks
- probe/probe_inner.rs: all unsafe FFmpeg calls behind pub(crate) probe_file()

Delete info.rs which mixed safe and unsafe code in a single 1889-line file.
No behavior or API changes.

Closes #707
