+++ crates/ff-probe/src/probe/mod.rs
@@ -1,0 +1,6 @@
+//! Probe module — media file metadata extraction.
+
+mod builder;
+pub(crate) mod probe_inner;
+
+pub use builder::open;
